COMP: fix comparison warning

+++ Modules/MitkExt/DataManagement/mitkPlanarFigure.cpp
@@ -104,7 +104,7 @@ bool mitk::PlanarFigure::SetControlPoint( unsigned int index, const Point2D& poi
 
 bool mitk::PlanarFigure::SetCurrentControlPoint( const Point2D& point )
 {
-  if ( (m_SelectedControlPoint < 0) || (m_SelectedControlPoint >= m_NumberOfControlPoints) )
+  if ( (m_SelectedControlPoint < 0) || (m_SelectedControlPoint >= (int)m_NumberOfControlPoints) )
   {
     return false;
   }
